Vil Flytte BarnBosted til egen fil
